+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/block/tile/BlockEntityFramingTable.java
@@ -110,11 +110,18 @@ public class BlockEntityFramingTable extends BaseBlockEntity implements Nameable
         if (blockItem.getBlock() instanceof IFramedBlock) {
             FrameData data = stack.getOrDefault(ModDataComponents.FRAME_DATA.get(), FrameData.EMPTY);
 
-            if (!data.side().isEmpty() && !materialData.getSide().isEmpty())
+            /*if (!data.side().isEmpty() && !materialData.getSide().isEmpty())
                 return false;
             if (!data.trim().isEmpty() && !materialData.getTrim().isEmpty())
                 return false;
             if (!data.front().isEmpty() && !materialData.getFront().isEmpty())
+                return false;*/
+
+            if (!testTargetFrameMaterial(data.side(), materialData.getSide(), stack.getCount(), inventory().getItem(SLOT_SIDE).getCount()))
+                return false;
+            if (!testTargetFrameMaterial(data.front(), materialData.getFront(), stack.getCount(), inventory().getItem(SLOT_FRONT).getCount()))
+                return false;
+            if (!testTargetFrameMaterial(data.trim(), materialData.getTrim(), stack.getCount(), inventory().getItem(SLOT_TRIM).getCount()))
                 return false;
 
             return true;
@@ -124,6 +131,18 @@ public class BlockEntityFramingTable extends BaseBlockEntity implements Nameable
             return true;
 
         return false;
+    }
+
+    private boolean testTargetFrameMaterial(ItemStack sourceMat, ItemStack targetMat, int sourceCount, int targetCount) {
+        if (sourceMat.isEmpty() || targetMat.isEmpty())
+            return true;
+
+        if (!ItemStack.isSameItemSameComponents(sourceMat, targetMat))
+            return false;
+
+        return true;
+        //int limit = Math.min(sourceMat.getMaxStackSize(), targetMat.getMaxStackSize());
+        //return (sourceCount + targetCount) <= limit;
     }
 
     public static boolean isItemValidMaterial (ItemStack stack) {