+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Possible fix for restoring detached drawer leaving drawer without a label
 - Prevent drawers with contents being able to stack if contents identical
 - Prevent drawers from being stored in bundles and shulker boxes
+- Added config option to prevent storing filled drawers in drawers
+- Added config option to blacklist items or namespaces from being stored
 - FABRIC: Item labels are now shaded
 - NEO/FORGE: Shift+clicking an upgrade on a drawer will store it